src/requirements.txt
scapy[basic]
awsiotsdk
cachetools
tldextract
requests
pytz
certifi==2020.11.8 